refactor: refactoring PostTemplate
