+++ src/templates/PostTemplate/constants/index.ts
@@ -1,0 +1,6 @@
+export const FORM_STATUS = {
+  IDLE: "idle",
+  SUBMITTING: "submitting",
+  SUCCESS: "success",
+  ERROR: "error",
+} as const;
