+++ src/templates/PostTemplate/types/index.ts
@@ -1,0 +1,3 @@
+import { FORM_STATUS } from "../constants";
+
+export type FormStatus = (typeof FORM_STATUS)[keyof typeof FORM_STATUS];
